poli

+++ src/exercicios/POO/polimorfismo/Aula012/Animal.java
@@ -2,7 +2,7 @@ package exercicios.POO.polimorfismo.Aula012;
 
 abstract class Animal {
 
-    protected float peso;
+    protected double peso;
     protected int idade;
     protected int membros;
 
@@ -10,7 +10,7 @@ abstract class Animal {
     public abstract void  alimentar();
     public abstract void emitirSom();
 
-    public float getPeso() {
+    public double getPeso() {
         return peso;
     }
